Initial import.
This DatabaseMessageSource is a first try to make a i18n like messagesource for the springframework. One that does not use a file based messagesource but a database based on.

It is not finished but it does work for 1 basic language as of now and it should not be to difficult to make it work like it should.